+++ rpmdeplint/cli.py
@@ -9,14 +9,12 @@ import logging
 import sys
 from collections.abc import Callable
 from enum import IntEnum
-from importlib import metadata
 
-from rpmdeplint import DependencyAnalyzer, UnreadablePackageError
+from rpmdeplint import __version__
+from rpmdeplint.analyzer import DependencyAnalyzer, UnreadablePackageError
 from rpmdeplint.repodata import PackageDownloadError, Repo, RepoDownloadError
 
 logger = logging.getLogger(__name__)
-
-version = metadata.version("rpmdeplint")
 
 
 class ExitCode(IntEnum):
@@ -143,13 +141,16 @@ def dependency_analyzer_from_args(args):
     return DependencyAnalyzer(repos, list(args.rpms), arch=args.arch)
 
 
-def comma_separated_repo(value: str) -> Repo:
-    if "," not in value:
-        raise argparse.ArgumentTypeError(
-            f"Repo {value!r} is not in the form <name>,<path>"
-        )
-    repo_name, baseurl = value.split(",", 1)
-    return Repo(repo_name, baseurl)
+def repo(value: str) -> Repo:
+    if "," in value:
+        repo_name, url_or_path = value.split(",", 1)
+        if url_or_path.startswith("http") and "/metalink?" in url_or_path:
+            return Repo(name=repo_name, metalink=url_or_path)
+        return Repo(name=repo_name, baseurl=url_or_path)
+
+    if repos := list(Repo.from_yum_config(name=value)):
+        return repos[0]
+    raise ValueError(f"Repo {value} is not configured")
 
 
 def add_common_dependency_analyzer_args(parser):
@@ -162,12 +163,12 @@ def add_common_dependency_analyzer_args(parser):
     parser.add_argument(
         "-r",
         "--repo",
-        metavar="NAME,URL",
-        type=comma_separated_repo,
+        metavar="NAME[,URL_OR_PATH]",
+        type=repo,
         action="append",
         dest="repos",
         default=[],
-        help="Name and URL of a repo to test against",
+        help="Name and optional (baseurl or metalink or local path) of a repo to test against",
     )
     parser.add_argument(
         "-R",
@@ -213,7 +214,9 @@ def main():
         "--debug", action="store_true", help="Show detailed progress messages"
     )
     parser.add_argument("--quiet", action="store_true", help="Show only errors")
-    parser.add_argument("--version", action="version", version=f"%(prog)s {version}")
+    parser.add_argument(
+        "--version", action="version", version=f"%(prog)s {__version__}"
+    )
 
     subparsers = parser.add_subparsers(dest="subcommand", title="subcommands")
     subparsers.required = True