+++ rpmdeplint/repodata.py
@@ -4,27 +4,26 @@
 # (at your option) any later version.
 
 
-import configparser
-import errno
-import glob
-import logging
 import os
-import shutil
-import tempfile
-import time
+from collections.abc import Iterator
+from configparser import ConfigParser
+from contextlib import suppress
+from logging import getLogger
 from os import getenv
 from pathlib import Path
-from typing import BinaryIO, Optional
+from tempfile import mkdtemp, mkstemp
+from time import time
+from typing import Any, BinaryIO, Optional, Union
 
 import librepo
-import requests
 import rpm
+from requests import Session
 
-logger = logging.getLogger(__name__)
-requests_session = requests.Session()
+logger = getLogger(__name__)
+requests_session = Session()
 
 
-REPO_CACHE_DIR = os.path.join(os.sep, "var", "tmp")
+REPO_CACHE_DIR = "/var/tmp"
 REPO_CACHE_NAME_PREFIX = "rpmdeplint-"
 
 
@@ -40,74 +39,87 @@ class RepoDownloadError(Exception):
     """
 
 
-def get_yumvars() -> dict[str, str]:
-    # This is not all the yumvars, but hopefully good enough...
+class Cache:
+    @staticmethod
+    def base_path() -> Path:
+        default_cache_home = Path.home() / ".cache"
+        return Path(getenv("XDG_CACHE_HOME", default_cache_home)) / "rpmdeplint"
 
-    try:
-        import dnf.conf
-        import dnf.rpm
-    except ImportError:
-        pass
-    else:
-        installroot = ""
-        subst = dnf.conf.Conf().substitutions
-        subst["releasever"] = dnf.rpm.detect_releasever(installroot)
-        return subst
+    @staticmethod
+    def entry_path(checksum: str) -> Path:
+        return Cache.base_path() / checksum[:1] / checksum[1:]
 
-    try:
-        import rpmUtils
-        import yum
-        import yum.config
-    except ImportError:
-        pass
-    else:
-        return {
-            "arch": rpmUtils.arch.getCanonArch(),
-            "basearch": rpmUtils.arch.getBaseArch(),
-            "releasever": yum.config._getsysver(
-                "/", ["system-release(releasever)", "redhat-release"]
-            ),
-        }
-
-    # Probably not going to work but there's not much else we can do...
-    return {
-        "arch": "$arch",
-        "basearch": "$basearch",
-        "releasever": "$releasever",
-    }
-
-
-def substitute_yumvars(s: str, yumvars: dict[str, str]) -> str:
-    for name, value in yumvars.items():
-        s = s.replace(f"${name}", value)
-    return s
-
-
-def cache_base_path() -> Path:
-    default_cache_home = Path.home() / ".cache"
-    return Path(getenv("XDG_CACHE_HOME", default_cache_home)) / "rpmdeplint"
-
-
-def cache_entry_path(checksum: str) -> Path:
-    return cache_base_path() / checksum[:1] / checksum[1:]
-
-
-def clean_cache():
-    expiry_time = time.time() - float(
-        os.environ.get("RPMDEPLINT_EXPIRY_SECONDS", "604800")
-    )
-    if not cache_base_path().is_dir():
-        return  # nothing to do
-    for subdir in cache_base_path().iterdir():
-        # Should be a subdirectory named after the first checksum letter
-        if not subdir.is_dir():
-            continue
-        for entry in subdir.iterdir():
-            if not entry.is_file():
+    @staticmethod
+    def clean():
+        expiry_time = time() - float(getenv("RPMDEPLINT_EXPIRY_SECONDS", 604800))
+        if not Cache.base_path().is_dir():
+            return  # nothing to do
+        for subdir in Cache.base_path().iterdir():
+            # Should be a subdirectory named after the first checksum letter
+            if not subdir.is_dir():
                 continue
-            if entry.stat().st_mtime < expiry_time:
-                logger.debug("Purging expired cache file %s", entry)
-                entry.unlink()
+            for entry in subdir.iterdir():
+                if not entry.is_file():
+                    continue
+                if entry.stat().st_mtime < expiry_time:
+                    logger.debug("Purging expired cache file %s", entry)
+                    entry.unlink()
+
+    @staticmethod
+    def download_repodata_file(checksum: str, urls: list[str]) -> BinaryIO:
+        """
+        Each created file in cache becomes immutable, and is referenced in
+        the directory tree within XDG_CACHE_HOME as
+        $XDG_CACHE_HOME/rpmdeplint/<checksum-first-letter>/<rest-of-checksum>
+
+        Both metadata and the files to be cached are written to a tempdir first
+        then renamed to the cache dir atomically to avoid them potentially being
+        accessed before written to cache.
+        """
+        if not urls:
+            raise ValueError("No urls specified to download repodata from")
+
+        filepath_in_cache: Path = Cache.entry_path(checksum)
+        if filepath_in_cache.is_file():
+            f = filepath_in_cache.open(mode="rb")
+            logger.debug("Using cached file %s for %s", filepath_in_cache, urls[0])
+            # Bump the modtime on the cache file we are using,
+            # since our cache expiry is LRU based on modtime.
+            os.utime(f.fileno())
+            return f
+
+        filepath_in_cache.parent.mkdir(parents=True, exist_ok=True)
+        fd, temp_path = mkstemp(dir=filepath_in_cache.parent, text=False)
+        try:
+            f = os.fdopen(fd, "wb+")
+        except Exception:
+            os.close(fd)
+            raise
+        try:
+            for index, url in enumerate(urls):
+                try:
+                    logger.debug("Downloading %s", url)
+                    response = requests_session.get(url, stream=True)
+                    response.raise_for_status()
+                    for chunk in response.raw.stream(decode_content=False):
+                        f.write(chunk)
+                    response.close()
+                    break
+                except OSError as e:
+                    msg = f"Failed to download repodata file {url}, {e}"
+                    if index == len(urls) - 1:  # last url in the list
+                        raise RepoDownloadError(msg) from e
+                    logger.debug(msg)
+            f.flush()
+            f.seek(0)
+            os.fchmod(f.fileno(), 0o644)
+            os.rename(temp_path, filepath_in_cache)
+            logger.debug("Cached as %s", filepath_in_cache)
+            return f
+        except Exception:
+            f.close()
+            os.unlink(temp_path)
+            raise
 
 
 class Repo:
@@ -115,20 +127,42 @@ class Repo:
     Represents a Yum ("repomd") package repository to test dependencies against.
     """
 
-    yum_main_config_path = "/etc/yum.conf"
-    yum_repos_config_glob = "/etc/yum.repos.d/*.repo"
+    yum_repos_d = Path("/etc/yum.repos.d/")
+
+    @staticmethod
+    def get_yumvars() -> dict[str, str]:
+        with suppress(ModuleNotFoundError):
+            import dnf.conf
+            import dnf.rpm
+
+            subst = dnf.conf.Conf().substitutions
+            subst["releasever"] = dnf.rpm.detect_releasever(installroot="")
+            return subst
+
+        # Probably not going to work but there's not much else we can do...
+        return {
+            "arch": "$arch",
+            "basearch": "$basearch",
+            "releasever": "$releasever",
+        }
 
     @classmethod
-    def from_yum_config(cls):
+    def from_yum_config(cls, name: str = "") -> Iterator["Repo"]:
         """
-        Yields Repo instances loaded from the system-wide Yum
-        configuration in :file:`/etc/yum.conf` and :file:`/etc/yum.repos.d/`.
+        Yields Repo instances loaded from the system-wide
+        configuration in :file:`/etc/yum.repos.d/`.
         """
-        yumvars = get_yumvars()
-        config = configparser.RawConfigParser()
-        config.read([cls.yum_main_config_path, *glob.glob(cls.yum_repos_config_glob)])
+
+        def substitute_yumvars(s: str, _yumvars: dict[str, str]) -> str:
+            for name, value in _yumvars.items():
+                s = s.replace(f"${name}", value)
+            return s
+
+        yumvars = cls.get_yumvars()
+        config = ConfigParser()
+        config.read(cls.yum_repos_d.glob("*.repo"))
         for section in config.sections():
-            if section == "main":
+            if name and section != name:
                 continue
             if config.has_option(section, "enabled") and not config.getboolean(
                 section, "enabled"
@@ -164,13 +198,13 @@ class Repo:
 
     def __init__(
         self,
-        repo_name: str,
+        name: str,
         baseurl: Optional[str] = None,
         metalink: Optional[str] = None,
         skip_if_unavailable: bool = False,
     ):
         """
-        :param repo_name: Name of the repository, for example "fedora-updates"
+        :param name: Name of the repository, for example "fedora-updates"
                           (used in problems and error messages)
         :param baseurl: URL or filesystem path to the base of the repository
                         (there should be a repodata subdirectory under this)
@@ -181,121 +215,79 @@ class Repo:
 
         Exactly one of the *baseurl* or *metalink* parameters must be supplied.
         """
-        self.name = repo_name
+        self.name = name
         if not baseurl and not metalink:
-            raise RuntimeError("Must specify either baseurl or metalink for repo")
-        self.baseurl = baseurl
+            raise ValueError("Must specify either baseurl or metalink for repo")
+        if baseurl and not baseurl.startswith("http"):
+            baseurl = baseurl.removeprefix("file://")
+            if not Path(baseurl).is_dir():
+                raise ValueError(f"baseurl {baseurl!r} is not a local directory")
+        self.urls = [baseurl] if baseurl else []
         self.metalink = metalink
         self.skip_if_unavailable = skip_if_unavailable
 
+        self.is_local = baseurl is not None and Path(baseurl).is_dir()
+        self._librepo_handle = librepo.Handle()
+        self._rpmmd_repomd: Optional[dict[str, Any]] = None
+        self.primary: Optional[BinaryIO] = None
+        self.filelists: Optional[BinaryIO] = None
+
     def download_repodata(self):
-        clean_cache()
-        logger.debug(
-            "Loading repodata for %s from %s", self.name, self.baseurl or self.metalink
-        )
-        self.librepo_handle = h = librepo.Handle()
-        r = librepo.Result()
-        h.repotype = librepo.LR_YUMREPO
-        if self.baseurl:
-            h.urls = [self.baseurl]
-        if self.metalink:
-            h.mirrorlist = self.metalink
-        h.setopt(
-            librepo.LRO_DESTDIR,
-            tempfile.mkdtemp(
-                self.name, prefix=REPO_CACHE_NAME_PREFIX, dir=REPO_CACHE_DIR
-            ),
-        )
-        h.setopt(librepo.LRO_INTERRUPTIBLE, True)
-        h.setopt(librepo.LRO_YUMDLIST, [])
-        if self.baseurl and os.path.isdir(self.baseurl):
-            self._download_metadata_result(h, r)
-            self._yum_repomd = r.yum_repomd
-            self._root_path = self.baseurl
-            self.primary = open(self.primary_url, "rb")
-            self.filelists = open(self.filelists_url, "rb")
+        Cache.clean()
+        self._download_metadata_result()
+        if self.is_local:
+            self.primary = open(self.primary_urls[0], "rb")
+            self.filelists = open(self.filelists_urls[0], "rb")
         else:
-            self._root_path = h.destdir = tempfile.mkdtemp(
-                self.name, prefix=REPO_CACHE_NAME_PREFIX, dir=REPO_CACHE_DIR
+            self.primary = Cache.download_repodata_file(
+                self.primary_checksum, self.primary_urls
             )
-            self._download_metadata_result(h, r)
-            self._yum_repomd = r.yum_repomd
-            self.primary = self._download_repodata_file(
-                self.primary_checksum, self.primary_url
-            )
-            self.filelists = self._download_repodata_file(
-                self.filelists_checksum, self.filelists_url
+            self.filelists = Cache.download_repodata_file(
+                self.filelists_checksum, self.filelists_urls
             )
 
-    def _download_metadata_result(self, handle: librepo.Handle, result: librepo.Result):
-        try:
-            handle.perform(result)
-        except librepo.LibrepoException as ex:
-            raise RepoDownloadError(
-                f"Failed to download repodata for {self!r}: {ex.args[1]}"
-            ) from ex
-
-    def _download_repodata_file(self, checksum: str, url: str) -> BinaryIO:
-        """
-        Each created file in cache becomes immutable, and is referenced in
-        the directory tree within XDG_CACHE_HOME as
-        $XDG_CACHE_HOME/rpmdeplint/<checksum-first-letter>/<rest-of-checksum>
-
-        Both metadata and the files to be cached are written to a tempdir first
-        then renamed to the cache dir atomically to avoid them potentially being
-        accessed before written to cache.
-        """
-        filepath_in_cache: Path = cache_entry_path(checksum)
-        try:
-            f = open(filepath_in_cache, "rb")
-        except OSError as e:
-            if e.errno == errno.ENOENT:
-                pass  # cache entry does not exist, we will download it
-            elif e.errno == errno.EISDIR:
-                # This is the original cache directory layout, merged in commit
-                # 6f11c3708, although it didn't appear in any released version
-                # of rpmdeplint. To be helpful we will fix it up, by just
-                # deleting the directory and letting it be replaced by a file.
-                shutil.rmtree(filepath_in_cache, ignore_errors=True)
-            else:
-                raise
-        else:
-            logger.debug("Using cached file %s for %s", filepath_in_cache, url)
-            # Bump the modtime on the cache file we are using,
-            # since our cache expiry is LRU based on modtime.
-            os.utime(f.fileno())  # Python 3.3+
-            return f
-        filepath_in_cache.parent.mkdir(parents=True, exist_ok=True)
-        fd, temp_path = tempfile.mkstemp(dir=filepath_in_cache.parent, text=False)
-        logger.debug("Downloading %s to cache temp file %s", url, temp_path)
-        try:
-            f = os.fdopen(fd, "wb+")
-        except Exception:
-            os.close(fd)
-            raise
-        try:
+    def _download_metadata_result(self) -> None:
+        def perform() -> librepo.Result:
             try:
-                response = requests_session.get(url, stream=True)
-                response.raise_for_status()
-                for chunk in response.raw.stream(decode_content=False):
-                    f.write(chunk)
-                response.close()
-            except OSError as e:
+                return self._librepo_handle.perform()
+            except librepo.LibrepoException as ex:
                 raise RepoDownloadError(
-                    f"Failed to download repodata file {os.path.basename(url)} for {self!r}: {e}"
-                ) from e
-            f.flush()
-            f.seek(0)
-            os.fchmod(f.fileno(), 0o644)
-            os.rename(temp_path, filepath_in_cache)
-            logger.debug("Using cached file %s for %s", filepath_in_cache, url)
-            return f
-        except Exception:
-            f.close()
-            os.unlink(temp_path)
-            raise
+                    f"Failed to download repo metadata for {self!r}: {ex.args[1]}"
+                ) from ex
 
-    def _is_header_complete(self, local_path: str) -> bool:
+        logger.debug(
+            "Loading repodata for repo %r from %s",
+            self.name,
+            self.urls or self.metalink,
+        )
+
+        h = self._librepo_handle
+        h.repotype = librepo.LR_YUMREPO
+        if self.urls:
+            h.urls = self.urls
+        if self.metalink:
+            h.metalinkurl = self.metalink
+        if self.is_local:
+            # no files will be downloaded
+            h.local = True
+        else:
+            # tempdir for repomd.xml (metadata) and downloaded rpms
+            h.destdir = mkdtemp(
+                self.name, prefix=REPO_CACHE_NAME_PREFIX, dir=REPO_CACHE_DIR
+            )
+        h.interruptible = True  # Download is interruptible
+        h.yumdlist = []  # Download only repomd.xml from repodata/
+
+        result = perform()
+        self._rpmmd_repomd = result.rpmmd_repomd
+
+        if self.metalink:
+            # Only download & parse metalink/mirrorlist
+            h.fetchmirrors = True
+            perform()
+            self.urls.extend(m for m in h.mirrors if m.startswith("http"))
+
+    def _is_header_complete(self, local_path: Union[Path, str]) -> bool:
         """
         Returns `True` if the RPM file `local_path` has complete RPM header.
         """
@@ -343,7 +335,7 @@ class Repo:
         only when complete RPM file is downloaded.
         """
         local_path = os.path.join(self._root_path, os.path.basename(location))
-        if self.librepo_handle.local:
+        if self.is_local:
             logger.debug("Using package %s from local filesystem directly", local_path)
             return local_path
 
@@ -358,7 +350,7 @@ class Repo:
                 location,
                 base_url=baseurl,
                 dest=self._root_path,
-                handle=self.librepo_handle,
+                handle=self._librepo_handle,
                 byterangestart=0,
                 byterangeend=byterangeend,
             )
@@ -379,36 +371,41 @@ class Repo:
         return target.local_path
 
     @property
-    def yum_repomd(self):
-        return self._yum_repomd
+    def _root_path(self) -> str:
+        # Path to the local repo dir or
+        # tempdir with downloaded repomd.xml and rpms
+        return self.urls[0] if self.is_local else self._librepo_handle.destdir
 
     @property
-    def repomd_fn(self) -> str:
-        return os.path.join(self._root_path, "repodata", "repomd.xml")
-
-    @property
-    def primary_url(self) -> str:
-        if not self.baseurl:
-            raise RuntimeError("baseurl not specified")
-        return os.path.join(self.baseurl, self.yum_repomd["primary"]["location_href"])
+    def repomd(self) -> dict[str, Any]:
+        if not self._rpmmd_repomd:
+            raise RuntimeError("_rpmmd_repomd is not set")
+        return self._rpmmd_repomd
 
     @property
     def primary_checksum(self) -> str:
-        return self.yum_repomd["primary"]["checksum"]
+        return self.repomd["records"]["primary"]["checksum"]
+
+    @property
+    def primary_urls(self) -> list[str]:
+        location_href = self.repomd["records"]["primary"]["location_href"]
+        return [os.path.join(url, location_href) for url in self.urls]
 
     @property
     def filelists_checksum(self) -> str:
-        return self.yum_repomd["filelists"]["checksum"]
+        return self.repomd["records"]["filelists"]["checksum"]
 
     @property
-    def filelists_url(self) -> str:
-        if not self.baseurl:
-            raise RuntimeError("baseurl not specified")
-        return os.path.join(self.baseurl, self.yum_repomd["filelists"]["location_href"])
+    def filelists_urls(self) -> list[str]:
+        location_href = self.repomd["records"]["filelists"]["location_href"]
+        return [os.path.join(url, location_href) for url in self.urls]
 
     def __repr__(self):
-        if self.baseurl:
-            return f"{self.__class__.__name__}(repo_name={self.name!r}, baseurl={self.baseurl!r})"
-        if self.metalink:
-            return f"{self.__class__.__name__}(repo_name={self.name!r}, metalink={self.metalink!r})"
-        return f"{self.__class__.__name__}(repo_name={self.name!r})"
+        return (
+            "Repo("
+            f"name={self.name!r}, "
+            f"urls={self.urls!r}, "
+            f"metalink={self.metalink!r}, "
+            f"skip_if_unavailable={self.skip_if_unavailable!r}, "
+            f"is_local={self.is_local!r})"
+        )